requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Existing dependencies (updated versions)
flask>=3.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
torch>=2.0.0
python-dotenv>=1.0.0

# Document processing
pdfminer.six>=20221105
python-docx>=0.8.11
markdown>=3.5.0

# Vector and search capabilities
faiss-cpu>=1.7.4
whoosh>=2.7.4

# LLM and embedding support
openai>=1.3.0
langchain>=0.0.340
langchain-community>=0.0.1

# File locking and utilities
filelock>=3.13.0

# Web scraping and text processing
beautifulsoup4>=4.12.0
lxml>=4.9.3

# JSON handling and validation
pydantic>=2.5.0

# Logging and monitoring
structlog>=23.2.0

# Additional utilities
requests>=2.31.0